+++ src/tap-bridge/model/tap-creator.cc
@@ -189,12 +189,7 @@ SendSocket(const char* path, int fd)
 }
 
 static int
-CreateTap(const char* dev,
-          const char* gw,
-          const char* ip,
-          const char* mac,
-          const char* mode,
-          const char* netmask)
+CreateTap(const char* dev, const char* ip, const char* mac, const char* mode, const char* netmask)
 {
     //
     // Creation and management of Tap devices is done via the tun device
@@ -275,7 +270,6 @@ main(int argc, char* argv[])
 {
     int c;
     char* dev = (char*)"";
-    char* gw = nullptr;
     char* ip = nullptr;
     char* mac = nullptr;
     char* netmask = nullptr;
@@ -284,15 +278,12 @@ main(int argc, char* argv[])
 
     opterr = 0;
 
-    while ((c = getopt(argc, argv, "vd:g:i:m:n:o:p:")) != -1)
+    while ((c = getopt(argc, argv, "vd:i:m:n:o:p:")) != -1)
     {
         switch (c)
         {
         case 'd':
             dev = optarg; // name of the new tap device
-            break;
-        case 'g':
-            gw = optarg; // gateway address for the new device
             break;
         case 'i':
             ip = optarg; // ip address of the new device
@@ -322,14 +313,6 @@ main(int argc, char* argv[])
     // create the device for us.  This name is given in dev
     //
     LOG("Provided Device Name is \"" << dev << "\"");
-
-    //
-    // We have got to be able to provide a gateway to the external Linux host
-    // so it can talk to the ns-3 network.  This ip address is provided in
-    // gw.
-    //
-    ABORT_IF(gw == nullptr, "Gateway Address is a required argument", 0);
-    LOG("Provided Gateway Address is \"" << gw << "\"");
 
     //
     // We have got to be able to assign an IP address to the tap device we are
@@ -385,7 +368,7 @@ main(int argc, char* argv[])
     // us to execute the following code:
     //
     LOG("Creating Tap");
-    int sock = CreateTap(dev, gw, ip, mac, operatingMode, netmask);
+    int sock = CreateTap(dev, ip, mac, operatingMode, netmask);
     ABORT_IF(sock == -1, "main(): Unable to create tap socket", 1);
 
     //